Move trace.h
