+++ Poisson_surface_reconstruction_3/include/CGAL/poisson_refine_triangulation.h
@@ -25,7 +25,7 @@
 
 
 // CGAL
-#include <CGAL/trace.h>
+#include <CGAL/IO/trace.h>
 #include <CGAL/Mesher_level.h>
 #include <CGAL/Mesh_3/Poisson_refine_cells_3.h>
 #include <CGAL/Poisson_mesh_cell_criteria_3.h>
