+++ Stream_support/include/CGAL/IO/trace.h
@@ -1,10 +1,9 @@
 // Copyright (c) 2007-09  INRIA Sophia-Antipolis (France).
 // All rights reserved.
-//
-// This file is part of CGAL (www.cgal.org).
-// You can redistribute it and/or modify it under the terms of the GNU
-// General Public License as published by the Free Software Foundation,
-// either version 3 of the License, or (at your option) any later version.
+// This file is part of CGAL (www.cgal.org); you can redistribute it and/or
+// modify it under the terms of the GNU Lesser General Public License as
+// published by the Free Software Foundation; either version 3 of the License,
+// or (at your option) any later version.
 //
 // Licensees holding a valid commercial license may use this file in
 // accordance with the commercial license agreement provided with the software.
@@ -14,15 +13,12 @@
 //
 // $URL$
 // $Id$
-// SPDX-License-Identifier: GPL-3.0+
+// SPDX-License-Identifier: LGPL-3.0+
 //
 // Author(s) : Laurent Saboret
 
-#ifndef CGAL_TRACE_H
-#define CGAL_TRACE_H
-
-#include <CGAL/license/Point_set_processing_3.h>
-
+#ifndef CGAL_IO_TRACE_H
+#define CGAL_IO_TRACE_H
 
 #include <stdio.h>
 #include <stdarg.h>
@@ -35,7 +31,7 @@
 // ---------------
 
 // print_stderr() = printf-like function to print to stderr
-inline void print_stderr(const char *fmt, ...)
+inline void CGAL_print_stderr(const char *fmt, ...)
 {
   va_list argp;
   va_start(argp, fmt);
@@ -46,9 +42,9 @@ inline void print_stderr(const char *fmt, ...)
 // CGAL_TRACE() = printf-like function to print to stderr
 // if DEBUG_TRACE is defined (ignored otherwise)
 #ifdef DEBUG_TRACE
-  #define CGAL_TRACE  print_stderr
+#define CGAL_TRACE  CGAL_print_stderr
 #else
-  #define CGAL_TRACE  if (false) print_stderr
+  #define CGAL_TRACE  if (false) CGAL_print_stderr
 #endif
 
 // CGAL_TRACE_STREAM = C++ stream that prints to std::cerr 
@@ -61,4 +57,4 @@ inline void print_stderr(const char *fmt, ...)
 
 /// \endcond
 
-#endif // CGAL_TRACE_H
+#endif // CGAL_IO_TRACE_H
